+++ Api/BitfinexApi/Json/SocketObjets/BitfinexSocketOrderBook.cs
@@ -12,10 +12,10 @@ namespace Bitfinex.Json.SocketObjets
     public class BitfinexSocketOrderBook
     {
         [BitfinexProperty(0)]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         [BitfinexProperty(1)]
         public int Count { get; set; }
         [BitfinexProperty(2)]
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
     }
 }
